feat(server-rs): session replay record, list, events and delete routes
Ports POST /api/session-replay/record/:siteId and the replay reads
(GET /api/sites/:siteId/session-replay/list, GET and DELETE
/api/sites/:siteId/session-replay/:sessionId) from recordSessionReplay.ts,
sessionReplayIngestService.ts, trackingUtils.ts, sessionReplayQueryService.ts
and the three read handlers.

- replay::json: an iterative JSON tape that reproduces
  JSON.stringify(JSON.parse(x)) byte for byte (lone surrogates kept as
  escapes, V8 number spelling, array-index key order, repeated keys),
  checked against 1500 Node-generated documents
- replay::schema: the zod schema with zod's issue order and shapes
- replay::ingest: fingerprint, identified user, Redis session, clock skew,
  session_replay_events and session_replay_metadata_v2 rows in Node's key
  order, UTF-16 sizes, luxon bound formatting in the process zone
- replay::record: Fastify body parsing, replay-off and unknown sites,
  client IP, Site Exclusion Decision, 500 bodies; usage gates always pass
  and R2 is disabled (no CLOUD)
- replay::query / read: Node's SQL text verbatim, publicReplayRead and
  authReplayWrite chains, traits enrichment, find-my-way quirks (empty
  params, DELETE .../list, maxParamLength, no Allow header)
- ports sessionReplayIngestService.test.ts and recordSessionReplay.test.ts
- analytics::chain: bind the segment site lookup as int8, matching auth's
  identical statement (sqlx's per-connection statement cache otherwise
  failed segment_id requests with a 500)
- parity/replay: differential harness against Node (responses, rows,
  reads with every credential kind, deletes, ClickHouse query log) and
  fixture generators

+++ server-rs/src/analytics/chain.rs
@@ -183,7 +183,9 @@ async fn expand_segment(
     // loadSegmentForSite
     let organization_id: Option<Option<String>> =
         sqlx::query_scalar("SELECT organization_id FROM sites WHERE site_id = $1 LIMIT 1")
-            .bind(numeric_site_id)
+            // i64 like auth's identical statement: sqlx caches prepared statements per
+            // connection by SQL text, and an int4 bound to the cached int8 one fails
+            .bind(i64::from(numeric_site_id))
             .fetch_optional(&state.pg)
             .await
             .map_err(internal_error)?;
